requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# fasttext/train.py: 6
PyYAML == 5.1

# fasttext/test.py: 1
# fasttext/train.py: 2
fasttext == 0.9.1

# bert/printcm.py: 2
# fasttext/printcm.py: 2
matplotlib == 2.1.1

# bert/printcm.py: 1
# bert/run_classifier.py: 29
# bert/test.py: 8
# fasttext/printcm.py: 1
numpy == 1.13.3

# bert/run_classifier.py: 34,35,36,37
# bert/test.py: 13,14,15,16
pytorch_pretrained_bert == 0.6.2

# bert/printcm.py: 4,5,6
# bert/test.py: 9,10,11
# fasttext/printcm.py: 4,5,6
# fasttext/test.py: 3,4
# fasttext/train.py: 3,4
scikit_learn == 0.20.2

# bert/test.py: 19
# fasttext/preprocess.py: 6
# fasttext/test.py: 6
# fasttext/train.py: 8
tabulate == 0.8.2

# bert/run_classifier.py: 31,32
# bert/test.py: 6
torch == 1.1.0

# bert/run_classifier.py: 27
# bert/test.py: 18
# fasttext/textcleaner.py: 7
tqdm == 4.29.1
